@@ -6,13 +6,13 @@
 
   <groupId>org.nicolie</groupId>
   <artifactId>TowersForPGM</artifactId>
-  <version>3.0.0</version>
+  <version>4.0.0</version>
   <packaging>jar</packaging>
 
   <name>TowersForPGM</name>
 
   <properties>
-    <java.version>1.8</java.version>
+    <java.version>21</java.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
 
