@@ -8,8 +8,8 @@ import org.bukkit.entity.Player;
 import org.nicolie.towersforpgm.draft.AvailablePlayers;
 import org.nicolie.towersforpgm.draft.Captains;
 import org.nicolie.towersforpgm.draft.Draft;
-import org.nicolie.towersforpgm.draft.PickInventory;
 import org.nicolie.towersforpgm.draft.Teams;
+import org.nicolie.towersforpgm.gui.Picks;
 import org.nicolie.towersforpgm.utils.ConfigManager;
 import org.nicolie.towersforpgm.utils.LanguageManager;
 import org.nicolie.towersforpgm.utils.SendMessage;
@@ -21,7 +21,7 @@ public class AddCommand implements CommandExecutor {
   private final AvailablePlayers availablePlayers;
   private final Captains captains;
   private final Teams teams;
-  private final PickInventory pickInventory;
+  private final Picks pickInventory;
   private final LanguageManager languageManager;
 
   public AddCommand(
@@ -29,7 +29,7 @@ public class AddCommand implements CommandExecutor {
       Captains captains,
       Teams teams,
       LanguageManager languageManager,
-      PickInventory pickInventory) {
+      Picks pickInventory) {
     this.availablePlayers = availablePlayers;
     this.captains = captains;
     this.teams = teams;
