@@ -7,12 +7,18 @@ public class Stats {
   private int assists;
   private double damageDone;
   private double damageTaken;
+  private double bowAccuracy;
   private int points;
   private int wins;
   private int games;
   private int winstreak;
+  private int maxWinstreak;
+  private int elo;
+  private int lastElo;
+  private int maxElo;
+  private int eloChange;
 
-  // Constructor
+  // Constructor completo
   public Stats(
       String username,
       int kills,
@@ -20,37 +26,55 @@ public class Stats {
       int assists,
       double damageDone,
       double damageTaken,
+      double bowAccuracy,
       int points,
       int wins,
       int games,
-      int winstreak) {
+      int winstreak,
+      int maxWinstreak,
+      int elo,
+      int lastElo,
+      int maxElo) {
     this.username = username;
     this.kills = kills;
     this.deaths = deaths;
     this.assists = assists;
     this.damageDone = damageDone;
     this.damageTaken = damageTaken;
+    this.bowAccuracy = bowAccuracy;
     this.points = points;
     this.wins = wins;
     this.games = games;
     this.winstreak = winstreak;
+    this.maxWinstreak = maxWinstreak;
+    this.elo = elo;
+    this.lastElo = lastElo;
+    this.maxElo = maxElo;
+    this.eloChange = elo - lastElo; // Calcular automáticamente el delta
   }
 
   @Override
   public String toString() {
-    return "PlayerStatUpdate{" + "username='"
+    return "Stats{" + "username='"
         + username + '\'' + ", kills="
         + kills + ", deaths="
         + deaths + ", assists="
         + assists + ", damageDone="
         + damageDone + ", damageTaken="
-        + damageTaken + ", points="
+        + damageTaken + ", bowAccuracy="
+        + bowAccuracy + ", points="
         + points + ", wins="
         + wins + ", games="
         + games + ", winstreak="
-        + winstreak + '}';
+        + winstreak + ", maxWinstreak="
+        + maxWinstreak + ", elo="
+        + elo + ", lastElo="
+        + lastElo + ", maxElo="
+        + maxElo + ", eloChange="
+        + eloChange + '}';
   }
 
+  // Getters
   public String getUsername() {
     return username;
   }
@@ -75,6 +99,10 @@ public class Stats {
     return damageTaken;
   }
 
+  public double getBowAccuracy() {
+    return bowAccuracy;
+  }
+
   public int getPoints() {
     return points;
   }
@@ -89,5 +117,34 @@ public class Stats {
 
   public int getWinstreak() {
     return winstreak;
+  }
+
+  public int getMaxWinstreak() {
+    return maxWinstreak;
+  }
+
+  public int getElo() {
+    return elo;
+  }
+
+  public int getLastElo() {
+    return lastElo;
+  }
+
+  public int getMaxElo() {
+    return maxElo;
+  }
+
+  public int getEloChange() {
+    return eloChange;
+  }
+
+  // Métodos calculados adicionales
+  public double getKDRatio() {
+    if (deaths == 0) {
+      return kills > 0 ? kills : 0.0;
+    } else {
+      return (double) kills / deaths;
+    }
   }
 }
